Restructured Profile page - created Profile card
