+++ client/src/components/ActiveRequestList.js
@@ -76,7 +76,6 @@ export default function ActiveRequestList() {
               {result}
           </Grid>
         </div>  
-        
 
     );
   }
